requirements.txt
Django==3.0.6
Pillow==7.1.2
sorl-thumbnail==12.6.3
